pyproject.toml
[project]
name = "synchronizer"
version = "0.1.0"
description = "Synchronize the teams and members from the `contributors/` and `teams/` directories to ScottyLabs services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "hvac>=2.4.0",
    "pydantic>=2.12.5",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "python-keycloak>=6.0.0",
    "slack-sdk>=3.39.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10",
    "ty>=0.0.10",
    "types-colorama>=0.4.15.20250801",
    "types-hvac>=2.4.0.20251115",
]

[project.scripts]
sync = "synchronizer:main"

[tool.setuptools.packages.find]
where = ["__meta/synchronizer"]
include = ["synchronizer*"]

[tool.mypy]
strict = true

# Supported rules https://github.com/charliermarsh/ruff/#supported-rules
[tool.ruff]
lint.select = ["ALL"]

lint.ignore = [
    # Do not require documentation for every single function/method.
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    # Ignore rules with lower code number when they are at conflict
    "D203",
    "D212",
    # May cause conflicts when used with the formatter
    "COM812",
]
